KelvinVR: Working library with LED methods
