+++ MetaTempDriver/Hardware/library/KelvinVR/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "jira-plugin.workingProject": ""
+}
