+++ MetaTempDriver/Hardware/library/KelvinVR/library.properties
@@ -1,0 +1,10 @@
+name=KelvinVR
+version=1.0.0
+author=Your Name
+maintainer=Your Name <<a href="mailto:your.email@example.com">your.email@example.com</a>>
+sentence=A brief description of your library.
+paragraph=A more detailed description of your library and its features.
+category=Other
+url=<a href="https://github.com/yourusername/KelvinVR">https://github.com/yourusername/KelvinVR</a>
+architectures=*
+includes=KelvinVR.h
