+++ MetaTempDriver/Hardware/library/KelvinVR/src/KelvinVR.cpp
@@ -12,32 +12,34 @@
 #define MIN_BRIGHTNESS 0
 #define MAX_BRIGHTNESS 255
 
-#include "KelvinVRESP32.hpp"
+#include <Arduino.h>
 
-void KelvinVRESP32::begin()
-{
-  // ESP32-specific initialization
-}
+// #include <hardware/esp32/KelvinVRESP32.hpp>
 
-void KelvinVRESP32::update()
-{
-  // ESP32-specific update logic
-}
+// void KelvinVRESP32::begin()
+// {
+//   // ESP32-specific initialization
+// }
+
+// void KelvinVRESP32::update()
+// {
+//   // ESP32-specific update logic
+// }
 
 // Implement other functions
 
 #include "KelvinVR.h"
 
-Kelvin::Kelvin(string modelName)
+Kelvin::Kelvin(const String &modelName)
 {
   _modelName = modelName;
 }
 
 void Kelvin::fullPinSetup()
 {
-  for (int i = 0; i < sizeof(pinsOutput); i++)
+  for (int i = 0; i < sizeof(this->_pinsOutput) / sizeof(this->_pinsOutput[0]); i++)
   {
-    pinMode(pinsOutput[i], OUTPUT);
+    pinMode(this->_pinsOutput[i], OUTPUT);
   }
 }
 
@@ -52,7 +54,7 @@ void Kelvin::blinkLed(int ledPin)
 void Kelvin::fadeEffect(int led)
 {
   // set the brightness of pin 9:
-  analogWrite(led, brightness);
+  analogWrite(led, _brightness);
 
   // change the brightness for next time through the loop:
   _brightness = _brightness + _fadeAmount;
@@ -75,6 +77,7 @@ void Kelvin::ledOn()
 void Kelvin::ledOff()
 {
   digitalWrite(ledPlus, LOW);
+  digitalWrite(ledMinus, LOW);
 }
 
 void Kelvin::blinkFront(int delayTime)
